# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TradeShield B2B escrow marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for the current user (as buyer or supplier)
      parameters:
        - name: role
          in: query
          required: false
          description: Filter to orders where the current user is this role
          schema:
            type: string
            enum: [buyer, supplier]
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OrderStatus"
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Buyer creates an order against a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order (pending_supplier_confirmation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid order (e.g. below MOQ, insufficient stock)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order (buyer, supplier, or admin only)
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "403":
          description: Not a party to this order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/accept:
    post:
      operationId: acceptOrder
      tags: [orders]
      summary: Supplier accepts the order
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order moved to awaiting_payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "409":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/reject:
    post:
      operationId: rejectOrder
      tags: [orders]
      summary: Supplier rejects the order
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectOrderInput"
      responses:
        "200":
          description: Order moved to expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "409":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/pay:
    post:
      operationId: payOrder
      tags: [orders]
      summary: Buyer initiates escrow payment (mock Moolre Collections charge)
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order moved to payment_processing; a mock webhook will resolve to in_escrow shortly after
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "409":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/ship:
    post:
      operationId: shipOrder
      tags: [orders]
      summary: Supplier marks the order as shipped; starts the 72h auto-release window
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order moved to shipped, auto_release_at set to now + 72h
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "409":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/confirm-receipt:
    post:
      operationId: confirmReceipt
      tags: [orders]
      summary: Buyer confirms receipt; initiates disbursement (async — completes via webhook/polling)
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmReceiptInput"
      responses:
        "200":
          description: Order moved to payout_processing; completed when disbursement webhook confirms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "409":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/dispute:
    post:
      operationId: raiseDispute
      tags: [disputes]
      summary: Buyer or supplier raises a dispute on an order
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeInput"
      responses:
        "201":
          description: Dispute created; order moved to disputed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
        "409":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/request-refund:
    post:
      operationId: requestRefund
      tags: [orders]
      summary: Buyer requests refund after no shipment update (in_escrow, 5+ days since payment)
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "201":
          description: Refund request created as dispute; order moved to disputed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
        "409":
          description: Invalid state or too early
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/dispute-reply:
    post:
      operationId: replyToDispute
      tags: [disputes]
      summary: Supplier right-of-reply on an open dispute
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeReplyInput"
      responses:
        "201":
          description: Reply recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeReply"
        "409":
          description: No open dispute or not supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/ratings:
    post:
      operationId: createRating
      tags: [ratings]
      summary: Rate the counterparty of a completed order
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingInput"
      responses:
        "201":
          description: Rating recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
        "409":
          description: Order not completed, or already rated by this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
...
    OrderId:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    OrderStatus:
      type: string
      enum:
        - pending_supplier_confirmation
        - awaiting_payment
        - payment_processing
        - in_escrow
        - shipped
        - payout_processing
        - completed
        - disputed
        - post_release_disputed
        - expired
        - rejected
        - payout_failed
    Order:
      type: object
      required:
        [
          id,
          buyerId,
          supplierId,
          productId,
          quantity,
          totalAmount,
          platformFee,
          status,
          createdAt,
        ]
      properties:
        id:
          type: integer
        buyerId:
          type: integer
        supplierId:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        totalAmount:
          type: string
        platformFee:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        createdAt:
          type: string
          format: date-time
        shippedAt:
          type: ["string", "null"]
          format: date-time
        autoReleaseAt:
          type: ["string", "null"]
          format: date-time
        deliveryLocation:
          type: ["string", "null"]
        preferredDeliveryDate:
          type: ["string", "null"]
          format: date-time
        rejectReason:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
          format: date-time
        autoReleaseReminderSent:
          type: boolean
        confirmPhotoUrl:
          type: ["string", "null"]
        productName:
          type: string
          description: Product display name (included when listing orders)
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            product:
              $ref: "#/components/schemas/Product"
            buyer:
              $ref: "#/components/schemas/User"
            supplier:
              $ref: "#/components/schemas/User"
            disputes:
              type: array
              items:
                $ref: "#/components/schemas/Dispute"
    OrderInput:
      type: object
      required: [productId, quantity, deliveryLocation, preferredDeliveryDate]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
        deliveryLocation:
          type: string
          minLength: 1
        preferredDeliveryDate:
          type: string
          format: date
    RejectOrderInput:
      type: object
      properties:
        reason:
          type: string
          minLength: 1
...
    DisputeCategory:
      type: string
      enum: [quality, non_delivery, quantity, other]
    DisputeInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
        category:
          $ref: "#/components/schemas/DisputeCategory"
        evidenceUrls:
          type: array
          items:
            type: string
...
    ConfirmReceiptInput:
      type: object
      properties:
        photoUrl:
          type: string
          description: Optional photo URL attached at receipt confirmation
...